Rename nifxml_doc
